Create Pascal’s Triangle
Given an integer N, return the first N rows of Pascal’s triangle.
In Pascal’s triangle, each number is the sum of the two numbers directly above it :
Input Format: N = 5
Result:
    1
   1 1
  1 2 1
 1 3 3 1
1 4 6 4 1
Explanation: There are 5 rows in the output matrix. Each row corresponds to each one of the rows in the image shown above.